Extracted functions from loaddata used by thredds_download that don't need numpy, xarray, etc.
